embedding-service/requirements.txt
fastapi
uvicorn[standard]
qdrant-client[grpc]
sentence-transformers
nltk
python-multipart
PyMuPDF
python-docx